file3/serializers.py
from rest_framework import serializers

from file3.models import Notebook


class NotebookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notebook
        fields = '__all__'

