requirements.txt
asgiref==3.4.1
Django==3.2.7
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
future==0.18.2
mysqlclient==2.0.3
Pillow==8.4.0
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1

